third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-padding-shorthand-expected.txt

Tests that the padding shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('padding') is '10px 5px 4px 3px'
PASS computedStyle.getPropertyValue('padding') is '64px 80px 96px 112px'
PASS computedStyle.getPropertyValue('padding') is '5px 6px 7px 8px'
PASS successfullyParsed is true

TEST COMPLETE
